+++ Finalraytracer/Utilities/Vector3D.h
@@ -4,15 +4,15 @@
 class Vector3D
 {
 public:
-	double x, y, z;
+	float x, y, z;
 public:
 	Vector3D(void);
-	Vector3D(double a);
-	Vector3D(double _x, double _y, double _z);
+	Vector3D(float a);
+	Vector3D(float _x, float _y, float _z);
 	Vector3D(const Vector3D& v);
 	~Vector3D(void);
 
-	double length(void);
+	float length(void);
 	void normalize(void);
 
 	inline Vector3D operator+(const Vector3D& vec)
@@ -25,12 +25,12 @@ public:
 		return (Vector3D(x - vec.x, y - vec.y, z - vec.z));
 	}
 
-	inline Vector3D operator*(const double a)
+	inline Vector3D operator*(const float a)
 	{
 		return (Vector3D(x * a, y * a, z * a));
 	}
 
-	inline Vector3D operator/(const double a)
+	inline Vector3D operator/(const float a)
 	{
 		return (Vector3D(x / a, y / a, z / a));
 	}
@@ -49,23 +49,23 @@ public:
 		return (*this);
 	}
 
-	inline Vector3D operator*=(const double a)
+	inline Vector3D operator*=(const float a)
 	{
 		x *= a; y *= a; z *= a;
 		return (*this);
 	}
 
-	inline Vector3D operator/=(const double a)
+	inline Vector3D operator/=(const float a)
 	{
 		x /= a; y /= a; z /= a;
 		return (*this);
 	}
 
-	inline double dotProduct(const Vector3D& vec)
+	inline float dotProduct(const Vector3D& vec)
 	{
 		return (x * vec.x + y * vec.y + z * vec.z);
 	}
-	inline double lenSquared(void)
+	inline float lenSquared(void)
 	{
 		return x * x + y * y + z * z;
 	}
